requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
langchain-openai==0.0.2
langchain-pinecone==0.0.1
langchain==0.0.335
pinecone-client==2.2.4
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
markdown==3.4.3 